correcciones sprites
se corrigio un error de carga y colocacion en sprites luego de cierta posicion horizontal, se agregaron las nuevas sprites del juego

+++ JuegoTivaC/Personaje.h
@@ -9,14 +9,14 @@ class Personaje{
 
   private:
     uint8_t HBposx, HBposy; //posicion en X y Y de la barra de salud
-    uint8_t Posx, Posy; //posicion en X y Y del personaje al jugar
+    int Posx, Posy; //posicion en X y Y del personaje al jugar
     uint8_t Numb; //para indicar el numero de jugador que es
     
   public:
     uint8_t pose; //que pose tomara
     short skinsel; //que skin se va a seleccionar
     uint8_t health; //vida del personaje
-    Personaje(uint8_t Numb, uint8_t HBposx, uint8_t HBposy, uint8_t Posx, uint8_t Posy);
+    Personaje(uint8_t Numb, uint8_t HBposx, uint8_t HBposy,unsigned int Posx,unsigned int Posy);
     void init(); //imprime la barra de salud
     void updateSp(); //actualizar el sprite del personaje para crear el movimiento
     void takedamage(uint8_t dano); //cuando el personaje recibe daño
